expose decompileApk function
