+++ src/core/apktool/DecompileResult.ts
@@ -1,0 +1,7 @@
+interface DecompileResult {
+    isSuccessful: boolean;
+    decompileFolderPath?: string;
+    manifestPath?: string;
+    error?: string;
+    apkToolYmlPath?: string;
+}
